+++ CHANGELOG.md
@@ -49,6 +49,7 @@ Change categories:
 - platform: Add functions to decode Redfish Action PDR (DSP0248 v1.3.0)
 - dsp: platform: Iterator macros for State Effecter PDR
 - base: Added PLDM Base MultipartSend support
+- dsp: platform: Add requester APIs for SetNumericSensorEnable
 
 ### Changed